Parquet import (#57)
* parquet to csv import

This implementation was created to replace the previous nyc_taxi dag, given the latest updates from the NYC TLC which replaced the .csv files for Parquet ones. At the moment, CrateDB does not support the import of Parquet files, so this solution was developed to circumvent it. 
The workflow has the following steps:
- detect the current date to format the file name
- download the Parquet file and then transform it to csv using parquet-tools
- upload it to S3
- run COPY FROM query to copy the content to the staging table
- copy from staging table to trips table
- delete staging content
- delete processed files